Apply review recommendation
