+++ decidim-elections/app/helpers/decidim/elections/admin/elections_helper.rb
@@ -5,6 +5,7 @@ module Decidim
     module Admin
       module ElectionsHelper
         include Decidim::ApplicationHelper
+        include Decidim::Elections::LabelHelper
 
         def census_count(election)
           election.census&.count(election).to_i
@@ -22,23 +23,6 @@ module Decidim
           return user unless election.census
 
           election.census.user_presenter.constantize.new(user)
-        end
-
-        def election_status_with_label(election)
-          css_class = case election.status
-                      when :ongoing
-                        "warning"
-                      when :finished
-                        "success"
-                      when :unpublished
-                        "alert"
-                      else
-                        "reverse"
-                      end
-
-          content_tag(:span,
-                      I18n.t("decidim.elections.status.#{election.status}"),
-                      class: "#{css_class} label")
         end
 
         def election_status_action_data(election)
